# swagger.yaml
openapi: 3.0.3
info:
  title: ScaleIT Bridge API
  description: |
    REST API for communicating with industrial scales (Dini Argeo, Rinstrum, etc.) via TCP/IP or Serial connections.
    
    ## Important Notes:
    
    - **CORS**: All endpoints support CORS automatically (allow any origin, method, and header)
    - **Mixed Content**: If accessing from HTTPS (e.g., Vercel), browsers may block HTTP connections. Use IP address or HTTPS tunnel.
    - **Base URL**: 
      - Local: `http://localhost:8080` or `http://127.0.0.1:8080`
      - Network: `http://YOUR_COMPUTER_IP:8080` (e.g., `http://192.168.1.100:8080`)
    
    ## Authentication:
    No authentication required for local installations.
    
  version: 0.1.1
  contact:
    name: ScaleIT Bridge Support
    url: https://github.com/tkogut/Bridge_ScaleCmd_Rust
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://127.0.0.1:8080
    description: Local development server (alternative)
  - url: http://{ip}:8080
    description: Network server (replace {ip} with your computer's IP address)
    variables:
      ip:
        default: 192.168.1.100
        description: Your computer's IP address on the local network
...
paths:
...
  /devices:
    get:
      tags:
        - Devices
      summary: List all devices
      description: |
        Get a list of all configured scale devices.
        
        Returns an array of tuples: `[device_id, device_name, device_model]`
      operationId: listDevices
      responses:
        '200':
          description: List of configured devices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicesResponse'
              examples:
                success:
                  summary: Devices list
                  value:
                    success: true
                    devices:
                      - ["scale1", "Main Scale", "Dini Argeo C320"]
                      - ["scale2", "Secondary Scale", "Rinstrum R320"]
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /scalecmd:
    post:
      tags:
        - Scale Commands
      summary: Execute scale command
      description: |
        Execute a command on a scale device.
        
        **Available commands:**
        - `readGross`: Read gross weight
        - `readNet`: Read net weight
        - `tare`: Tare the scale
        - `zero`: Zero the scale
        
        **Timeout:** Commands have a configurable timeout (default: 5000ms). If the device doesn't respond within the timeout, an error is returned.
      operationId: executeScaleCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaleCommandRequest'
            examples:
              readGross:
                summary: Read gross weight
                value:
                  device_id: "scale1"
                  command: "readGross"
              readNet:
                summary: Read net weight
                value:
                  device_id: "scale1"
                  command: "readNet"
              tare:
                summary: Tare scale
                value:
                  device_id: "scale1"
                  command: "tare"
              zero:
                summary: Zero scale
                value:
                  device_id: "scale1"
                  command: "zero"
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScaleCommandResponse'
              examples:
                success:
                  summary: Successful weight reading
                  value:
                    success: true
                    device_id: "scale1"
                    command: "readGross"
                    result:
                      gross_weight: 123.45
                      net_weight: 123.45
                      unit: "kg"
                      is_stable: true
                      timestamp: "2025-12-19T10:30:00Z"
                error:
                  summary: Command failed
                  value:
                    success: false
                    device_id: "scale1"
                    command: "readGross"
                    result: null
                    error: "Connection timeout to device"
        '400':
          description: Bad request (e.g., empty device_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyDeviceId:
                  summary: Empty device_id
                  value:
                    success: false
                    error: "device_id is required"
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScaleCommandResponse'
              examples:
                timeout:
                  summary: Connection timeout
                  value:
                    success: false
                    device_id: "scale1"
                    command: "readGross"
                    result: null
                    error: "Connection timeout to 192.168.1.254:4001"
  /api/config:
    get:
      tags:
        - Configuration
      summary: Get all device configurations
      description: |
        Get detailed configuration for all devices.
        
        Returns a map where keys are device IDs and values are device configurations.
      operationId: getAllDeviceConfigs
      responses:
        '200':
          description: Device configurations
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/DeviceConfig'
              examples:
                configs:
                  summary: Device configurations
                  value:
                    scale1:
                      name: "Main Scale"
                      manufacturer: "Dini Argeo"
                      model: "C320"
                      host_id: "host-1"
                      miernik_id: "miernik-1"
                      enabled: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/save:
    post:
      tags:
        - Configuration
      summary: Save device configuration
      description: |
        Add or update a device configuration.
        
        After saving, the configuration is automatically reloaded and the device connection is re-established.
      operationId: saveDeviceConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveConfigRequest'
            examples:
              tcpDevice:
                summary: TCP device configuration
                value:
                  device_id: "scale1"
                  config:
                    name: "Main Scale"
                    manufacturer: "Dini Argeo"
                    model: "C320"
                    host_id: "host-1"
                    miernik_id: "miernik-1"
                    enabled: true
              deviceWithSerial:
                summary: Device with serial connection
                value:
                  device_id: "scale2"
                  config:
                    name: "Serial Scale"
                    manufacturer: "Rinstrum"
                    model: "R320"
                    host_id: "host-2"
                    miernik_id: "miernik-2"
                    enabled: true
      responses:
        '200':
          description: Configuration saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveConfigResponse'
              examples:
                success:
                  summary: Configuration saved
                  value:
                    success: true
                    message: "Configuration for scale1 saved and reloaded."
        '400':
          description: Bad request (invalid configuration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/config/{device_id}:
    delete:
      tags:
        - Configuration
      summary: Delete device configuration
      description: |
        Remove a device configuration.
        
        After deletion, the configuration is automatically reloaded.
      operationId: deleteDeviceConfig
      parameters:
        - name: device_id
          in: path
          required: true
          description: Device ID to delete
          schema:
            type: string
          examples:
            scale1:
              summary: Delete scale1
              value: "scale1"
      responses:
        '200':
          description: Device deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteConfigResponse'
              examples:
                success:
                  summary: Device deleted
                  value:
                    success: true
                    message: "Device scale1 deleted and configuration reloaded."
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/hosts:
    get:
      tags:
        - Hosts
      summary: List all host configurations
      description: |
        Get all host (connection) configurations.
        
        Returns a map where keys are host IDs and values are host configurations.
      operationId: getAllHosts
      responses:
        '200':
          description: Host configurations
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/HostConfig'
              examples:
                hosts:
                  summary: Host configurations
                  value:
                    host-1:
                      name: "Host 192.168.1.254:4001"
                      connection:
                        connection_type: "Tcp"
                        host: "192.168.1.254"
                        port: 4001
                      timeout_ms: 5000
                      enabled: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DevicesResponse:
      type: object
      required:
        - success
        - devices
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        devices:
          type: array
          description: List of devices as tuples [device_id, name, model]
          items:
            type: array
            minItems: 3
            maxItems: 3
            items:
              type: string
          example:
            - ["scale1", "Main Scale", "Dini Argeo C320"]
            - ["scale2", "Secondary Scale", "Rinstrum R320"]
    ScaleCommandRequest:
      type: object
      required:
        - device_id
        - command
      properties:
        device_id:
          type: string
          description: Device identifier
          example: "scale1"
        command:
          type: string
          enum: [readGross, readNet, tare, zero]
          description: Command to execute
          example: "readGross"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Device not found"
        message:
          type: string
          description: Alternative error message field
          example: "Connection timeout to device"